# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PerioVision AI API",
    "version": "2.0",
    "description": "Research prototype for periodontal decision support. Every JSON response uses the envelope {data, meta, error, mode}. Authenticate with POST /api/auth/login, then send 'Authorization: Bearer <access_token>'. x-permission names the RBAC permission each operation requires (see docs/SECURITY.md)."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Envelope": {
        "type": "object",
        "required": [
          "data",
          "meta",
          "error",
          "mode"
        ],
        "properties": {
          "data": {
            "description": "Payload (object, list or null on error)"
          },
          "meta": {
            "type": "object",
            "description": "Extra information such as counts"
          },
          "error": {
            "nullable": true,
            "type": "object",
            "properties": {
              "code": {
                "type": "integer"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            }
          },
          "mode": {
            "type": "string",
            "enum": [
              "demo",
              "live"
            ]
          }
        }
      },
...
      "UserCreateIn": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "email": {
            "maxLength": 254,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 256,
            "minLength": 10,
            "title": "Password",
            "type": "string"
          },
          "role": {
            "enum": [
              "admin",
              "dentist",
              "technician",
              "auditor"
            ],
            "title": "Role",
            "type": "string"
          },
          "clinic_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Clinic Name"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "role"
        ],
        "title": "UserCreateIn",
        "type": "object"
      },
      "UserUpdateIn": {
        "additionalProperties": false,
        "properties": {
          "role": {
            "anyOf": [
              {
                "enum": [
                  "admin",
                  "dentist",
                  "technician",
                  "auditor"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Role"
          },
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Active"
          }
        },
        "title": "UserUpdateIn",
        "type": "object"
      }
    }
  },
  "paths": {
    "/api/admin/config": {
      "get": {
        "summary": "Thresholds, key IDs and signing status",
        "operationId": "security_get_config_get",
        "tags": [
          "security"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "admin:config"
      }
    },
    "/api/admin/users": {
      "get": {
        "summary": "List user accounts",
        "operationId": "security_list_users_get",
        "tags": [
          "security"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "admin:users"
      },
      "post": {
        "summary": "Create a user account",
        "operationId": "security_create_user_post",
        "tags": [
          "security"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          },
          "422": {
            "description": "Validation failed; error.details lists each field"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "admin:users",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateIn"
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{user_id}": {
      "patch": {
        "summary": "Change a user's role or disable them",
        "operationId": "security_update_user_patch",
        "tags": [
          "security"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          },
          "422": {
            "description": "Validation failed; error.details lists each field"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "admin:users",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateIn"
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{user_id}/lock": {
      "post": {
        "summary": "Lock a user for a number of minutes",
        "operationId": "security_lock_user_post",
        "tags": [
          "security"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "admin:users",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/models/status": {
      "get": {
        "summary": "Model files and signature status",
        "operationId": "security_models_status_get",
        "tags": [
          "security"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "model:read"
      }
    },
    "/api/models/trust": {
      "get": {
        "summary": "Calibration, flagged share and model integrity",
        "operationId": "security_model_trust_get",
        "tags": [
          "security"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "model:read"
      }
    },
...
    "/api/security-lab": {
      "get": {
        "summary": "List Security Lab attack simulations",
        "operationId": "security_lab_scenarios_get",
        "tags": [
          "security"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "security_lab:run"
      }
    },
    "/api/security-lab/run-all": {
      "post": {
        "summary": "Run every attack simulation",
        "operationId": "security_lab_run_all_post",
        "tags": [
          "security"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "security_lab:run"
      }
    },
    "/api/security-lab/{scenario_id}/run": {
      "post": {
        "summary": "Run one attack simulation",
        "operationId": "security_lab_run_post",
        "tags": [
          "security"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "security_lab:run",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/security/rbac-matrix": {
      "get": {
        "summary": "Role permission matrix",
        "operationId": "security_rbac_matrix_get",
        "tags": [
          "security"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "self:manage"
      }
    },
    "/api/security/sessions": {
      "get": {
        "summary": "Active sessions of all users",
        "operationId": "security_all_sessions_get",
        "tags": [
          "security"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "security:read"
      }
    },
...
  }
}